requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-mistralai
